requirements.txt
tensorflow
fastapi
uvicorn
aiofiles
music21
matplotlib
midi2audio
python-decouple